# services/todo-mcp/handlers/handlers.go
package handlers

import (
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/modelcontextprotocol/go-sdk/mcp"
)

func Tools() gin.HandlerFunc {
	return func(c *gin.Context) {
		tools := []gin.H{
			{
				"id":          "add_todo",
				"name":        "add_todo",
				"description": "A tool to add a new todo item",
				"inputSchema": gin.H{
					"type": "object",
					"properties": gin.H{
						"title": gin.H{
							"type":        "string",
							"description": "The title of the todo item",
						},
						"due_date": gin.H{
							"type":        "string",
							"format":      "date-time",
							"description": "The due date of the todo item (optional)",
						},
					},
					"required": []string{"title"},
				},
			},
			{
				"id":          "get_todos",
				"name":        "get_todos",
				"description": "A tool to retrieve all todo items",
				"inputSchema": gin.H{
					"type":       "object",
					"properties": gin.H{},
				},
			},
			{
				"id":          "update_todo",
				"name":        "update_todo",
				"description": "A tool to update an existing todo item by ID",
				"inputSchema": gin.H{
					"type": "object",
					"properties": gin.H{
						"id": gin.H{
							"type":        "integer",
							"description": "The ID of the todo item to update",
						},
						"title": gin.H{
							"type":        "string",
							"description": "The new title for the todo item (optional)",
						},
						"due_date": gin.H{
							"type":        "string",
							"format":      "date-time",
							"description": "The new due date for the todo item (optional)",
						},
					},
					"required": []string{"id"},
				},
			},
			{
				"id":          "delete_todo",
				"name":        "delete_todo",
				"description": "A tool to delete a todo item by ID",
				"inputSchema": gin.H{
					"type": "object",
					"properties": gin.H{
						"id": gin.H{
							"type":        "integer",
							"description": "The ID of the todo item to delete",
						},
					},
					"required": []string{"id"},
				},
			},
		}

		c.JSON(http.StatusOK, gin.H{
			"tools": tools,
		})
	}
}

// InvokeTool returns a handler that invokes a specific tool
// Note: This is a REST-like endpoint for testing. For production MCP usage, use the /sse endpoint
func InvokeTool() gin.HandlerFunc {
	return func(c *gin.Context) {
		toolID := c.Param("id")

		// Parse request body
		var requestBody struct {
			Arguments map[string]interface{} `json:"arguments"`
		}

		if err := c.ShouldBindJSON(&requestBody); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid request body"})
			return
		}

		// Execute the tool directly (bypassing MCP protocol for HTTP convenience)
		var result *mcp.CallToolResult
		var err error

		switch toolID {
		case "add_todo":
			title, _ := requestBody.Arguments["title"].(string)
			dueDate, _ := requestBody.Arguments["due_date"].(string)
			input := CreateTodoInput{
				Title:   title,
				DueDate: dueDate,
			}
			result, _, err = CreateTodo(c.Request.Context(), nil, input)

		case "get_todos":
			input := GetTodosInput{}
			result, _, err = GetTodos(c.Request.Context(), nil, input)

		case "update_todo":
			// Handle ID as either float64 (from JSON) or int
			var id int32
			switch v := requestBody.Arguments["id"].(type) {
			case float64:
				id = int32(v)
			case int:
				id = int32(v)
			case int32:
				id = v
			}

			title, _ := requestBody.Arguments["title"].(string)
			dueDate, _ := requestBody.Arguments["due_date"].(string)
			input := UpdateTodoInput{
				ID:      id,
				Title:   title,
				DueDate: dueDate,
			}
			result, _, err = UpdateTodo(c.Request.Context(), nil, input)

		case "delete_todo":
			// Handle ID as either float64 (from JSON) or int
			var id int32
			switch v := requestBody.Arguments["id"].(type) {
			case float64:
				id = int32(v)
			case int:
				id = int32(v)
			case int32:
				id = v
			}

			input := DeleteTodoInput{ID: id}
			result, _, err = DeleteTodo(c.Request.Context(), nil, input)

		default:
			c.JSON(http.StatusNotFound, gin.H{"error": fmt.Sprintf("Tool '%s' not found", toolID)})
			return
		}

		if err != nil {
			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
			return
		}

		// Build response
		response := gin.H{
			"content": result.Content,
		}

		if result.IsError {
			response["isError"] = true
		}

		c.JSON(http.StatusOK, response)
	}
}

// schemaHandler returns the MCP protocol schema
func Schema() gin.HandlerFunc {
	return func(c *gin.Context) {
		schema := gin.H{
			"version": "1.0.0",
			"protocol": gin.H{
				"name":    "Model Context Protocol",
				"version": "2024-11-05",
			},
			"server": gin.H{
				"name":    "todo-mcp",
				"version": "1.0.0",
			},
		}

		c.JSON(http.StatusOK, schema)
	}
}

// CapabilitiesHandler returns the server capabilities
func Capabilities() gin.HandlerFunc {
	return func(c *gin.Context) {
		capabilities := gin.H{
			"tools": gin.H{
				"supported": true,
			},
			"prompts": gin.H{
				"supported": false,
			},
			"resources": gin.H{
				"supported": false,
			},
		}

		c.JSON(http.StatusOK, capabilities)
	}
}