# services/todo-mcp/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Todo MCP Server API",
    "description": "REST API for the Todo MCP server providing tool invocation endpoints for testing. For production MCP usage, use the SSE endpoint.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/tools/get_todos/invoke": {
      "post": {
        "summary": "Get all todo items",
        "description": "Retrieves all todo items from the todo list",
        "operationId": "getTodos",
        "tags": ["Tools"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "arguments": {
                    "type": "object",
                    "properties": {},
                    "description": "No arguments required for this tool"
                  }
                },
                "required": ["arguments"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List of todos retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/tools/update_todo/invoke": {
      "post": {
        "summary": "Update a todo item",
        "description": "Updates an existing todo item by ID",
        "operationId": "updateTodo",
        "tags": ["Tools"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "arguments": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "format": "int32",
                        "description": "The ID of the todo item to update",
                        "example": 1
                      },
                      "title": {
                        "type": "string",
                        "description": "The new title for the todo item (optional)",
                        "example": "Buy groceries and cook dinner"
                      },
                      "due_date": {
                        "type": "string",
                        "format": "date-time",
                        "description": "The new due date for the todo item (optional)",
                        "example": "2024-12-31T23:59:59Z"
                      }
                    },
                    "required": ["id"]
                  }
                },
                "required": ["arguments"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Todo item updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Todo item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResult"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/tools/delete_todo/invoke": {
      "post": {
        "summary": "Delete a todo item",
        "description": "Deletes a todo item by ID",
        "operationId": "deleteTodo",
        "tags": ["Tools"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "arguments": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "format": "int32",
                        "description": "The ID of the todo item to delete",
                        "example": 1
                      }
                    },
                    "required": ["id"]
                  }
                },
                "required": ["arguments"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Todo item deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Todo item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResult"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ToolResult": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "text"
                },
                "text": {
                  "type": "string",
                  "description": "The result text from the tool execution"
                }
              }
            }
          },
          "isError": {
            "type": "boolean",
            "description": "Indicates if the tool execution resulted in an error"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message",
            "example": "Invalid request body"
          }
        }
      }
    }
  },
...
}